+++ examples/ng/doc/toh-6/lib/src/dashboard_component.dart
@@ -13,7 +13,7 @@ import 'hero_search_component.dart';
   selector: 'my-dashboard',
   templateUrl: 'dashboard_component.html',
   styleUrls: const ['dashboard_component.css'],
-  directives: const [CORE_DIRECTIVES, HeroSearchComponent, ROUTER_DIRECTIVES],
+  directives: const [CORE_DIRECTIVES, HeroSearchComponent, routerDirectives],
 )
 // #enddocregion search
 class DashboardComponent implements OnInit {