+++ examples/ng/doc/toh-6/lib/src/hero_search_component.dart
@@ -50,11 +50,5 @@ class HeroSearchComponent implements OnInit {
   }
   // #enddocregion search
 
-  void gotoDetail(Hero hero) {
-    var link = [
-      'HeroDetail',
-      {'id': hero.id.toString()}
-    ];
-    _router.navigate(link);
-  }
+  Future gotoDetail(Hero hero) => _router.navigate('/detail/${hero.id}');
 }